feat(Application, GameManager, InputValidator): add GameManager
Application:
- call GameManager activate function
GameManager:
- get input from user
- set game properties
- manage game
InputValidator:
- change class name CarValidator to InputValidator
- validate car name input
- validate game cnt input
README:
- edit function list
