+++ README.md
@@ -15,6 +15,7 @@
 - 예외 상황에 대한 에러문구 출력하기
     - 자동차 이름은 5자 이내로 입력하기 
     - 시도 횟수는 숫자로 입력하기
+- (GameManager) GameManager 로 게임에 대한 관리 책임 이전
